Move all the selection-specific variable code into the `selection` module
