+++ apollo-federation/src/sources/connect/validation/mod.rs
@@ -22,7 +22,6 @@ mod http;
 mod link;
 mod schema;
 mod source;
-mod variable;
 
 use std::fmt::Display;
 use std::ops::Range;
