--- apollo-federation/src/sources/connect/validation/variable/resolver/this.rs
@@ -1,60 +1,0 @@
-use apollo_compiler::ast::FieldDefinition;
-use apollo_compiler::schema::Component;
-use apollo_compiler::schema::ExtendedType;
-
-use crate::sources::connect::validation::Code;
-use crate::sources::connect::validation::Message;
-use crate::sources::connect::validation::graphql::GraphQLString;
-use crate::sources::connect::validation::graphql::SchemaInfo;
-use crate::sources::connect::validation::variable::resolver;
-use crate::sources::connect::validation::variable::resolver::NamespaceResolver;
-use crate::sources::connect::variable::Namespace;
-use crate::sources::connect::variable::VariableReference;
-
-/// Resolves variables in the `$this` namespace
-pub(crate) struct ThisResolver<'a> {
-    object: &'a ExtendedType,
-    field: &'a Component<FieldDefinition>,
-}
-
-impl<'a> ThisResolver<'a> {
-    pub(crate) fn new(object: &'a ExtendedType, field: &'a Component<FieldDefinition>) -> Self {
-        Self { object, field }
-    }
-}
-
-impl NamespaceResolver for ThisResolver<'_> {
-    fn check(
-        &self,
-        reference: &VariableReference<Namespace>,
-        expression: GraphQLString,
-        schema: &SchemaInfo,
-    ) -> Result<(), Message> {
-        let Some(root) = resolver::get_root(reference) else {
-            return Ok(()); // Not something we can type check this way
-        };
-
-        let fields = match self.object {
-            ExtendedType::Object(node) => &node.fields,
-            _ => return Ok(()), // TODO: interfaces
-        };
-
-        let field_type = fields
-            .get(root.as_str())
-            .ok_or_else(|| Message {
-                code: Code::UndefinedField,
-                message: format!(
-                    "`{object}` does not have a field named `{root}`",
-                    object = self.object.name(),
-                    root = root.as_str(),
-                ),
-                locations: expression
-                    .line_col_for_subslice(root.location.start..root.location.end, schema)
-                    .into_iter()
-                    .collect(),
-            })
-            .map(|field| field.ty.clone())?;
-
-        resolver::resolve_path(schema, reference, expression, &field_type, self.field)
-    }
-}
